Timemachine.Data (timemachine v1.0.1)

The Data context.

Functions

clock_create(attrs \\ %{}, user_id)

Creates a clock for the given user_id. attrs should contains "status" and "time".

clock_get_last_by_user_id(user_id)

Returns the last clock made by the user user_id.

clock_list_by_user_id(user_id)

Returns all clocks made by the user user_id.

team_create(attrs \\ %{})

Creates a new team. attrs should contain "name".

team_delete(team)

Deletes the given team.

team_get(id)

Returns the team with the same id.

team_list()

Returns all teams.

team_update(team, attrs)

Edits the team with the given attrs. attrs can contain "name".

user_create(attrs \\ %{})

Creates a new user. attrs should contain "username", "email", "password" and a optional "role".

user_delete(user)

Deletes the given user.

user_get(id)

Returns the team with the same id.

user_search(params)

Returns the bearer token from a conn by reading authorization header.

user_try_login(credentials)

Tries to authenticate the user. credentials is a Map with keys "username" and "password".

user_try_login(user_id, password)

Returns the bearer token from a conn by reading authorization header.

user_update(user, attrs)

Returns the bearer token from a conn by reading authorization header.

userteam_create(user_id, team_id)

Returns the bearer token from a conn by reading authorization header.

userteam_delete(user_id, team_id)

Returns the bearer token from a conn by reading authorization header.

workingtime_create(attrs \\ %{}, user_id)

Returns the bearer token from a conn by reading authorization header.

workingtime_delete(workingtime)

Returns the bearer token from a conn by reading authorization header.

workingtime_generate(user_id, end_time)

Returns the bearer token from a conn by reading authorization header.

workingtime_get(id)

Returns the workingtime with the same id.

workingtime_list()

Returns the bearer token from a conn by reading authorization header.

workingtime_search(user_id, start_working, end_working)

Returns the bearer token from a conn by reading authorization header.

workingtime_update(workingtime, attrs)

Returns the bearer token from a conn by reading authorization header.