TimemachineWeb.Plugs.CSRFValidate (timemachine v1.0.1)

Plug for CSRF token validation.

Description

This plug is responsible for validating CSRF tokens for POST, PUT, and DELETE requests. It checks the presence of a CSRF token in the request and verifies its validity. If the token is invalid or missing, it falls back to the FallbackController with a 401 Unauthorized error. If the token is present but not authorized for the current user, it falls back with a 403 Forbidden error.

Usage

Add this plug to your pipeline in the router or controller to enforce CSRF token validation.

pipeline :api do
  plug TimemachineWeb.Plugs.CSRFValidate
  # ... other plugs and middleware
end

Functions

call(conn, default)

Call function for the CSRFValidate plug.

Parameters conn (type: Plug.Conn.t()): The connection struct. _default (type: term()): The default value. Returns If CSRF token is valid, the connection is returned. If the token is invalid or missing, it falls back to the FallbackController with a 401 Unauthorized error. If the token is present but not authorized for the current user, it falls back with a 403 Forbidden error.

init(default)

Callback implementation for Plug.init/1.