TimemachineWeb.UserController (timemachine v1.0.1)

Controller for managing user data.

Actions

search(conn, params): Search for users based on specified parameters.

login(conn, %{"credentials" => credentials}): Login using username and password, sends JWT and CSRF tokens.

tokens(conn, _params): Send JWT and CSRF tokens for the current user.

create(conn, %{"user" => user_params}): Create a new user.

show(conn, %{"id" => id}): Retrieve details of a specific user.

update(conn, %{"id" => id, "user" => user_params}): Update details of a specific user.

update_password(conn, %{"id" => id, "password" => password, "new_password" => new_password}): Update user password.

delete(conn, %{"id" => id}): Delete a specific user.

Helper Functions

send_tokens(conn, user): Send JWT and CSRF tokens in the response headers.

Functions

create(conn, map)

Create a new user.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "user".

Returns

Renders the :show view with the details of the created user.

delete(conn, map)

Delete a specific user.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "id".

Returns

Responds with a :no_content status.

login(conn, map)

Login using username and password, sends JWT and CSRF tokens.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "credentials".

Returns

Sends JWT and CSRF tokens in the response headers.

search(conn, params)

Search for users based on specified parameters.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing search parameters.

Returns

Renders the :search view with a list of users based on search parameters.

show(conn, map)

Retrieve details of a specific user.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "id".

Returns

Renders the :show view with details of the specified user.

tokens(conn, params)

Send JWT and CSRF tokens for the current user.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters (not used).

Returns

Sends JWT and CSRF tokens in the response headers.

update(conn, map)

Update details of a specific user.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "id" and "user".

Returns

Renders the :show view with the updated details of the user.

update_password(conn, map)

Update user password.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "id", "password", and "new_password".

Returns

Sends JWT and CSRF tokens in the response headers.