TimemachineWeb.WorkingtimeController (timemachine v1.0.1)

Controller for managing working times.

Actions

index(conn, params): Get a list of working times based on user_id, start_time, and end_time.

create(conn, %{"user_id" => user_id, "workingtime" => workingtime_params}): Create a working time entry.

show(conn, %{"id" => id, "user_id" => user_id}): Show details of a working time entry.

update(conn, %{"id" => id, "workingtime" => workingtime_params}): Update a working time entry.

delete(conn, %{"id" => id}): Delete a working time entry.

get_by_team(conn, params): Get a list of working times for a team based on team_id, start_time, and end_time.

create_for_team(conn, %{"team_id" => team_id, "workingtime" => workingtime_params}): Create working time entries for a team.

Functions

create(conn, map)

Create a working time entry.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "user_id" and "workingtime".

Returns

Responds with the created working time entry.

create_for_team(conn, map)

Create working time entries for a team.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "team_id" and "workingtime".

Returns

Responds with a :no_content status.

delete(conn, map)

Delete a working time entry.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "id".

Returns

Responds with a :no_content status.

get_by_team(conn, params)

Get a list of working times for a team based on team_id, start_time, and end_time.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "team_id", "start", and "end".

Returns

Responds with a list of working times for the specified team.

index(conn, params)

Get a list of working times based on user_id, start_time, and end_time.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "user_id", "start", and "end".

Returns

Responds with a list of working times.

show(conn, map)

Show details of a working time entry.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "id" and "user_id".

Returns

Responds with the details of the specified working time entry.

update(conn, map)

Update a working time entry.

Parameters

conn (type: Plug.Conn.t()): The connection struct.

params (type: map()): Map containing parameters, including "id" and "workingtime".

Returns

Responds with the updated working time entry.