Timemachine.Data (timemachine v1.0.1)

The Data context for managing data operations in the Timemachine application.

Functions

clock_create(attrs \\ %{}, user_id)

Creates a clock entry for the given user_id.

Parameters

attrs (type: map, default: %{}): Map containing attributes for the clock entry. Should include "status" and "time".

user_id (type: integer()): The user ID for whom the clock entry is created.

Returns

Returns :ok if the clock entry is successfully created.

clock_get_last_by_user_id(user_id)

Returns the last clock entry made by the user with the given user_id.

Parameters

user_id (type: integer()): The user ID for whom the last clock entry is retrieved.

Returns

Returns the last clock entry as a %Clock{} struct.

clock_list_by_user_id(user_id)

Returns all clock entries made by the user with the given user_id.

Parameters

user_id (type: integer()): The user ID for whom the clock entries are retrieved.

Returns

Returns a list of clock entries as %Clock{} structs.

team_create(attrs \\ %{})

Creates a new team.

Parameters

attrs (type: map, default: %{}): Map containing attributes for the team. Should include "name".

Returns

Returns :ok if the team is successfully created.

team_delete(team)

Deletes the given team.

Parameters

team (type: %Team{}): The team to be deleted.

Returns

Returns :ok if the team is successfully deleted.

team_get(id)

Returns the team with the same id.

Parameters

id (type: any()): The ID of the team to be retrieved.

Returns

Returns the team as a %Team{} struct.

team_list()

Returns all teams.

Returns

Returns a list of teams as %Team{} structs.

team_update(team, attrs)

Edits the team with the given attrs.

Parameters

team (type: %Team{}): The team to be updated.

attrs (type: map()): Map containing attributes for the team. Can include "name".

Returns

Returns :ok if the team is successfully updated.

user_create(attrs \\ %{})

Creates a new user.

Parameters

attrs (type: map()): Map containing user attributes, including "username", "email", "password", and an optional "role".

Returns

Returns the newly created user, preloaded with associated teams.

user_delete(user)

Deletes the given user.

Parameters

user (type: %User{}): The user to be deleted.

Returns

Returns :ok if the user is successfully deleted.

user_get(id)

Returns the team with the same id.

Parameters

id (type: any()): The ID of the user to be retrieved.

Returns

Returns the user as a %User{} struct.

user_search(params)

Searches for users based on the specified parameters.

Parameters

params (type: map()): Map containing search parameters. Can include "username" and "email".

Returns

Returns a list of users as %User{} structs.

user_try_login(credentials)

Tries to authenticate the user.

Parameters

credentials (type: map()): Map containing credentials with keys "username" and "password".

Returns

Returns {:ok, user} if authentication is successful, or {:error, 401} if not.

user_try_login(user_id, password)

Tries to authenticate the user.

Parameters

user_id (type: any()): The ID of the user.

password (type: string()): The user's password.

Returns

Returns {:ok, user} if authentication is successful, or {:error, 401} if not.

user_update(user, attrs)

Updates the user's information.

Parameters

user (type: %User{}): The user to be updated.

attrs (type: map()): Map containing attributes to be updated.

Returns

Returns :ok if the user is successfully updated.

userteam_create(user_id, team_id)

Creates a new user-team relationship.

Parameters

user_id (type: any()): The ID of the user.

team_id (type: any()): The ID of the team.

Returns

Returns :ok if the user-team relationship is successfully created.

userteam_delete(user_id, team_id)

Deletes an existing user-team relationship.

Parameters

user_id (type: any()): The ID of the user.

team_id (type: any()): The ID of the team.

Returns

Returns :ok if the user-team relationship is successfully deleted.

workingtime_create(attrs \\ %{}, user_id)

Creates a new working time entry.

Parameters

attrs (type: map, default: %{}): Map containing attributes for the working time. Should include "start" and "end".

user_id (type: any()): The ID of the user.

Returns

Returns :ok if the working time entry is successfully created.

workingtime_delete(workingtime)

Deletes an existing working time entry.

Parameters

workingtime (type: %Workingtime{}): The working time entry to be deleted.

Returns

Returns :ok if the working time entry is successfully deleted.

workingtime_generate(user_id, end_time)

Generates a new working time entry based on the last clock entry.

Parameters

user_id (type: any()): The ID of the user.

end_time (type: any()): The end time of the working time.

Returns

Returns :ok if the working time entry is successfully generated.

workingtime_get(id)

Returns the working time entry with the same id.

Parameters

id (type: any()): The ID of the working time entry.

Returns

Returns the working time entry as a %Workingtime{} struct.

workingtime_list()

Returns a list of all working time entries.

Returns

Returns a list of working time entries as %Workingtime{} structs.

workingtime_search(user_id, start_working, end_working)

Searches for working time entries within a specified time range.

Parameters

user_id (type: any()): The ID of the user.

start_working (type: any()): The start of the time range for the working time entries.

end_working (type: any()): The end of the time range for the working time entries.

Returns

Returns a list of working time entries within the specified time range.

workingtime_update(workingtime, attrs)

Updates an existing working time entry.

Parameters

workingtime (type: %Workingtime{}): The working time entry to be updated.

attrs (type: map()): Map containing attributes to be updated.

Returns

Returns :ok if the working time entry is successfully updated.